# docs/swagger.yaml
basePath: /api
definitions:
  api_err.ApiErr:
    properties:
      msg:
        type: string
      typ:
        $ref: '#/definitions/api_err.ApiErrTyp'
    type: object
  api_err.ApiErrTyp:
    enum:
    - not_found
    - bad_input
    - forbidden
    - unauthorized
    - already_claimed
    - pending_claim_exists
    - username_invalid
    - username_taken
    - report_target_missing
    - report_locked
    - duplicate_open_report
    - text_too_long
    - unknown
    type: string
    x-enum-varnames:
    - AetNotFound
    - AetBadInput
    - AetForbidden
    - AetUnauthorized
    - AetAlreadyClaimed
    - AetPendingClaimExists
    - AetUsernameInvalid
    - AetUsernameTaken
    - AetReportTargetMissing
    - AetReportLocked
    - AetDuplicateOpenReport
    - AetTextTooLong
    - AetUnknown
  models.Admin:
    properties:
      avatar_key:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      name:
        type: string
      power:
        type: number
      username:
        type: string
    type: object
...
  models.Category:
    properties:
      created_at:
        type: string
      desc:
        $ref: '#/definitions/models.I18nText'
      emoji:
        type: string
      id:
        type: string
      name:
        $ref: '#/definitions/models.I18nText'
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  models.HourRange:
    properties:
      close:
        description: '"22:00"'
        type: string
      open:
        description: '"09:00"'
        type: string
    type: object
  models.I18nText:
    properties:
      en:
        type: string
      uz:
        type: string
    type: object
...
  models.Status:
    enum:
    - -10
    - 0
    - 10
    type: integer
    x-enum-varnames:
    - StatusRejected
    - StatusPending
    - StatusApproved
...
  models.WeeklyHours:
    properties:
      fri:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      mon:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      sat:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      sun:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      thu:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      tue:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      wed:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
    type: object
  services.AdminEditPlaceInput:
    properties:
      address:
        $ref: '#/definitions/models.I18nText'
      category_id:
        type: string
      description:
        $ref: '#/definitions/models.I18nText'
      images:
        items:
          type: string
        type: array
      lat:
        type: number
      logo_key:
        type: string
      lon:
        type: number
      name:
        type: string
      phone:
        type: string
      weekly_hours:
        $ref: '#/definitions/models.WeeklyHours'
    type: object
  services.AdminLoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  services.AdminLoginOutput:
    properties:
      admin:
        $ref: '#/definitions/models.Admin'
      token:
        type: string
    type: object
  services.AdminUserDetailView:
    properties:
      avatar_key:
        type: string
      blocked:
        type: boolean
      claim_count:
        type: integer
      claimed_place_count:
        type: integer
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      place_count:
        type: integer
      report_count:
        type: integer
      review_count:
        type: integer
      telegram_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  services.BlockUserInput:
    properties:
      blocked:
        type: boolean
    type: object
...
  services.CreateAdminInput:
    properties:
      name:
        type: string
      password:
        type: string
      power:
        type: number
      username:
        type: string
    type: object
...
  services.EditAdminInput:
    properties:
      name:
        type: string
      password:
        type: string
      power:
        type: number
    type: object
  services.EditCategoryInput:
    properties:
      desc:
        $ref: '#/definitions/models.I18nText'
      emoji:
        type: string
      name:
        $ref: '#/definitions/models.I18nText'
    type: object
...
  services.Ok:
    properties:
      blocked:
        type: boolean
      ok:
        type: boolean
      status:
        $ref: '#/definitions/models.Status'
    type: object
...
  services.ReviewClaimInput:
    properties:
      status:
        $ref: '#/definitions/models.Status'
    type: object
...
  services.SetPlaceStatusInput:
    properties:
      status:
        $ref: '#/definitions/models.Status'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: S101 backend API
  title: S101 API
  version: "1.0"
paths:
  /admin/admins:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: list admins'
      tags:
      - admin-admins
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreateAdminInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: create admin'
      tags:
      - admin-admins
  /admin/admins/{id}:
    get:
      parameters:
      - description: Admin UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: get admin by ID'
      tags:
      - admin-admins
    put:
      consumes:
      - application/json
      parameters:
      - description: Admin UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.EditAdminInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: edit admin'
      tags:
      - admin-admins
  /admin/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AdminLoginOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      summary: Admin login
      tags:
      - admin-auth
  /admin/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Get current admin
      tags:
      - admin-auth
  /admin/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      security:
      - BearerAuth: []
      summary: 'Admin: list categories'
      tags:
      - admin-categories
  /admin/categories/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.EditCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: edit category'
      tags:
      - admin-categories
  /admin/claims:
    get:
      parameters:
      - description: pending|approved|rejected
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: list claims'
      tags:
      - admin-claims
  /admin/claims/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Decision
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ReviewClaimInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: approve or reject a claim'
      tags:
      - admin-claims
  /admin/places:
    get:
      parameters:
      - description: pending|approved|rejected
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: list places'
      tags:
      - admin-places
  /admin/places/{id}:
    delete:
      parameters:
      - description: Place UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: delete place'
      tags:
      - admin-places
    put:
      consumes:
      - application/json
      parameters:
      - description: Place UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.AdminEditPlaceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: edit place'
      tags:
      - admin-places
  /admin/places/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Place UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.SetPlaceStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: set place status'
      tags:
      - admin-places
...
  /admin/reviews:
    get:
      parameters:
      - description: Filter by place ID
        in: query
        name: place_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: list reviews'
      tags:
      - admin-reviews
  /admin/reviews/{id}:
    delete:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: delete review'
      tags:
      - admin-reviews
  /admin/users:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: list users'
      tags:
      - admin-users
  /admin/users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AdminUserDetailView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: get user by ID'
      tags:
      - admin-users
  /admin/users/{id}/block:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Block flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.BlockUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.Ok'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: 'Admin: block/unblock user'
      tags:
      - admin-users
...
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"